# swagger.json
{
	"swagger": "2.0",
	"info": {
		"title": "Blogify API",
		"description": "Blogify app is a blog app",
		"version": "1.0",
		"contact": {
			"name": "Ant1g3n",
			"url": "https://github.com/ant1g3n",
			"email": "razaqayomide01@gmail.com"
		},
		"license": {
			"name": "MIT"
		},
		"termsOfService": "This API is free to consume by any application developer"
	},
	"host": "localhost:5000",
	"basePath": "/v1/api",
	"schemes": ["http", "https"],
	"consumes": ["application/x-www-form-urlencoded", "application/json"],
	"produces": ["application/json"],
...
	"paths": {
...
		"/auth/user/delete": {
			"delete": {
				"tags": ["Auth"],
				"summary": "Users can decide to delete their account",
				"description": "This endpoint uses delete request to delete a user's account",
				"operationId": "deleteUser",
				"consumes": ["application/json"],
				"produces": ["application/json"],
				"responses": {
					"500": {
						"description": "Internal server error / general error"
					},
					"401": {
						"description": "Invalid token, please authenticate"
					},
					"200": {
						"description": "User successfully deleted"
					}
				},
				"security": [
					{
						"JWT": []
					}
				]
			}
		}
	},
	"securityDefinitions": {
		"JWT": {
			"type": "apiKey",
			"name": "authorization",
			"in": "header"
		}
	}
}